python/requirements.txt
google-shopping-merchant-accounts==1.0.0
google-shopping-merchant-datasources==1.0.0
google-shopping-merchant-inventories==1.0.0
google-shopping-merchant-notifications==1.0.0
google-shopping-merchant-ordertracking==1.0.0
google-shopping-merchant-products==1.0.0
google-shopping-merchant-productstudio==0.1.1
google-shopping-merchant-promotions==1.0.0
google-shopping-merchant-quota==1.0.0
google-shopping-merchant-reports==1.0.0
google-shopping-merchant-reviews==0.2.0
google-auth-oauthlib
google-auth
googleapis-common-protos